chore: refactor public component wrappers for auto-answering into preview `Node` (#4108)
