+++ editor.planx.uk/src/@planx/components/Checklist/model.ts
@@ -1,7 +1,12 @@
 import { array } from "yup";
 
 import { BaseNodeData, Option } from "../shared";
-import { ChecklistLayout } from "./Public/components/VisibleChecklist";
+
+export enum ChecklistLayout {
+  Basic,
+  Grouped,
+  Images,
+}
 
 export interface Group<T> {
   title: string;
@@ -23,6 +28,7 @@ export interface Checklist extends BaseNodeData {
   allRequired?: boolean;
   categories?: Array<Category>;
   neverAutoAnswer?: boolean;
+  autoAnswers?: string[] | undefined;
 }
 
 interface ChecklistExpandableProps {
@@ -31,7 +37,7 @@ interface ChecklistExpandableProps {
 }
 
 export const toggleExpandableChecklist = (
-  checklist: ChecklistExpandableProps
+  checklist: ChecklistExpandableProps,
 ): ChecklistExpandableProps => {
   if (checklist.options !== undefined && checklist.options.length > 0) {
     return {
