+++ editor.planx.uk/src/@planx/components/Filter/Public.tsx
@@ -1,4 +1,3 @@
-import { useStore } from "pages/FlowEditor/lib/store";
 import { useEffect } from "react";
 
 import { PublicProps } from "../shared/types";
@@ -8,17 +7,12 @@ export type Props = PublicProps<Filter>;
 
 // Filters are always auto-answered and never seen by a user, but should still leave a breadcrumb
 export default function Component(props: Props) {
-  const autoAnswerableFlag = useStore((state) => state.autoAnswerableFlag);
-
-  let idThatCanBeAutoAnswered: string | undefined;
-  if (props.id) idThatCanBeAutoAnswered = autoAnswerableFlag(props.id);
-
   useEffect(() => {
     props.handleSubmit?.({
-      answers: [idThatCanBeAutoAnswered],
+      answers: [props.autoAnswer],
       auto: true,
     });
-  }, []);
+  }, [props.autoAnswer]);
 
   return null;
 }
