--- editor.planx.uk/src/@planx/components/Question/Public/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "Question.tsx"
-}
